localbazar/requirements.txt
Django==5.2.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
psycopg2-binary==2.9.9
supabase==2.3.0
python-dotenv==1.0.0
Pillow==10.1.0
django-filter==23.5
djangorestframework-simplejwt==5.3.0 